# docs/swagger/generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Convo API",
    "version": "1.0.0",
    "description": "\n      **Convo API Documentation**\n      \n      A comprehensive API for managing conversational forms, workspaces, and user interactions.\n      \n      ## Authentication\n      This API uses Clerk for authentication. Include your session token in the Authorization header:\n      `Authorization: Bearer <your-token>`\n      \n      ## Rate Limiting\n      API requests are rate limited per user:\n      - **Free Plan**: 100 requests/hour\n      - **Pro Plan**: 1000 requests/hour\n      - **Enterprise**: Custom limits\n      \n      Rate limit headers are included in all responses:\n      - `X-RateLimit-Limit`: Maximum requests allowed\n      - `X-RateLimit-Remaining`: Requests remaining in current window\n      - `X-RateLimit-Reset`: Time when the limit resets (Unix timestamp)\n    ",
    "contact": {
      "name": "API Support",
      "email": "support@convo.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-generated-at": "2025-07-01T10:52:11.739Z",
    "x-generator": "convo-forms-swagger-generator"
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    },
    {
      "url": "https://convo.ai/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ClerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token for user authentication"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for programmatic access"
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "message": {
                  "type": "string",
                  "example": "Authentication required to access this resource"
                },
                "code": {
                  "type": "string",
                  "example": "AUTH_REQUIRED"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Access forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                },
                "message": {
                  "type": "string",
                  "example": "Insufficient permissions to access this resource"
                },
                "code": {
                  "type": "string",
                  "example": "INSUFFICIENT_PERMISSIONS"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Not Found"
                },
                "message": {
                  "type": "string",
                  "example": "The requested resource was not found"
                },
                "code": {
                  "type": "string",
                  "example": "NOT_FOUND"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Validation Error"
                },
                "message": {
                  "type": "string",
                  "example": "Request validation failed"
                },
                "code": {
                  "type": "string",
                  "example": "VALIDATION_ERROR"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string",
                        "example": "email"
                      },
                      "message": {
                        "type": "string",
                        "example": "Invalid email format"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Internal Server Error"
                },
                "message": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "paths": {
...
    "/api/docs": {
      "get": {
        "summary": "Get OpenAPI specification",
        "description": "Returns the complete OpenAPI specification for the Convo Forms API",
        "tags": ["Documentation"],
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": ["json", "yaml"],
              "default": "json"
            },
            "description": "Response format (json or yaml)"
          },
          {
            "in": "query",
            "name": "version",
            "schema": {
              "type": "string",
              "default": "latest"
            },
            "description": "API version to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OpenAPI specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "OpenAPI 3.0 specification object"
                }
              },
              "application/x-yaml": {
                "schema": {
                  "type": "string",
                  "description": "OpenAPI specification in YAML format"
                }
              }
            }
          },
          "404": {
            "description": "API documentation not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "API documentation not found. Please run 'npm run docs:generate' to generate documentation."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Regenerate API documentation",
        "description": "Triggers regeneration of the OpenAPI specification",
        "tags": ["Documentation"],
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Documentation regenerated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "API documentation regenerated successfully"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "routesCount": {
                          "type": "integer",
                          "example": 25
                        },
                        "schemasCount": {
                          "type": "integer",
                          "example": 15
                        },
                        "generatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/forms/{id}/save-as-template": {
      "post": {
        "summary": "Save form as template",
        "description": "Creates a new template based on an existing form's configuration.\nThe form's config field is converted to the template's formSchema format.\nRequires create_template permission in the form's workspace.\n",
        "tags": ["Forms", "Templates"],
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Form ID to save as template"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "description": "Template name"
                  },
                  "description": {
                    "type": "string",
                    "description": "Template description"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "HR",
                      "Marketing",
                      "Support",
                      "Sales",
                      "Education",
                      "Healthcare",
                      "Other"
                    ],
                    "description": "Business category"
                  },
                  "thumbnailUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "Template preview image"
                  }
                }
              },
              "examples": {
                "saveAsTemplate": {
                  "summary": "Save customer survey form as template",
                  "value": {
                    "name": "Customer Satisfaction Survey Template",
                    "description": "Reusable template for customer satisfaction surveys",
                    "category": "Marketing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created successfully from form",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "template": {
                      "$ref": "#/components/schemas/Template"
                    },
                    "form": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "title": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Template created successfully from form"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}